groupchat/requirements.txt
requests>=2.31
python-dotenv>=1.0
pyaudio>=0.2.13
faster-whisper>=1.0
